# contracts/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "love-space API 契约（项目级 living 文件）",
    "description": "登记 OpenSpec change 触及的接口；operation 通过 x-requirement 反链到 openspec specs 的 {domain}/{Requirement 名}。存量未被 change 触及的接口暂不回溯登记。",
    "version": "2026.08.16"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "love-space-admin（JWT 认证，/api/admin/*）"
    },
    {
      "url": "http://localhost:8081",
      "description": "love-space-app（API-key 认证，/api/app/*）"
    }
  ],
  "paths": {
...
    "/api/admin/article-categories": {
      "get": {
        "summary": "栏目列表（sortOrder 升序）",
        "x-requirement": "article/文章栏目管理"
      },
      "post": {
        "summary": "创建栏目",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleCategoryUpsertRequest"
              }
            }
          }
        },
        "x-requirement": "article/文章栏目管理"
      }
    },
    "/api/admin/article-categories/{id}": {
      "put": {
        "summary": "更新栏目",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleCategoryUpsertRequest"
              }
            }
          }
        },
        "x-requirement": "article/文章栏目管理"
      },
      "delete": {
        "summary": "删除栏目（物理删除，不回写文章 categoryIds）",
        "x-requirement": "article/文章栏目管理"
      }
    },
    "/api/admin/articles/page": {
      "get": {
        "summary": "文章分页",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "x-requirement": "article/文章管理"
      }
    },
    "/api/admin/articles": {
      "post": {
        "summary": "创建文章（contentHtml 内 img src 存 objectKey，保存时逐个 validateAndBind）",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleUpsertRequest"
              }
            }
          }
        },
        "x-requirement": "article/文章管理"
      }
    },
    "/api/admin/articles/{id}": {
      "get": {
        "summary": "文章详情（contentHtml 读时签名替换）",
        "x-requirement": "article/文章管理"
      },
      "put": {
        "summary": "更新文章",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleUpsertRequest"
              }
            }
          }
        },
        "x-requirement": "article/文章管理"
      },
      "delete": {
        "summary": "删除文章（物理删除）",
        "x-requirement": "article/文章管理"
      }
    },
...
    "/api/admin/featured-items/page": {
      "get": {
        "summary": "精选推荐分页",
        "parameters": [
          {
            "name": "cityId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "x-requirement": "featured/精选推荐管理"
      }
    },
    "/api/admin/featured-items": {
      "post": {
        "summary": "创建精选推荐（cityId 必选且创建后不可变）",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeaturedItemUpsertRequest"
              }
            }
          }
        },
        "x-requirement": "featured/精选推荐管理"
      }
    },
    "/api/admin/featured-items/{id}": {
      "get": {
        "summary": "精选推荐详情",
        "x-requirement": "featured/精选推荐管理"
      },
      "put": {
        "summary": "更新精选推荐（忽略 cityId 变更）",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeaturedItemUpsertRequest"
              }
            }
          }
        },
        "x-requirement": "featured/精选推荐管理"
      },
      "delete": {
        "summary": "删除精选推荐（物理删除）",
        "x-requirement": "featured/精选推荐管理"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArticleCategoryUpsertRequest": {
        "type": "object",
        "required": [
          "name",
          "icon"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "description": "objectKey"
          },
          "sortOrder": {
            "type": "integer"
          }
        }
      },
      "ArticleUpsertRequest": {
        "type": "object",
        "required": [
          "image",
          "title"
        ],
        "properties": {
          "image": {
            "type": "string",
            "description": "objectKey"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "contentHtml": {
            "type": "string",
            "description": "富文本 HTML，img src 为 objectKey"
          },
          "sortOrder": {
            "type": "integer"
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "online": {
            "type": "boolean"
          }
        }
      },
      "FeaturedItemUpsertRequest": {
        "type": "object",
        "required": [
          "cityId",
          "banner"
        ],
        "properties": {
          "cityId": {
            "type": "string",
            "format": "uuid"
          },
          "banner": {
            "type": "string",
            "description": "objectKey，比例不校验"
          },
          "description": {
            "type": "string"
          },
          "online": {
            "type": "boolean"
          }
        }
      }
    }
  }
}